Split citation-evidence traversal under the graph structure caps for ticket 1145
